backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
strands-agents
strands-agents-tools
